完工

+++ src/main/java/czt/yonyou/dao/BillRepository.java
@@ -9,5 +9,5 @@ package czt.yonyou.dao;
 import czt.yonyou.entity.Bill;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BillRepository extends JpaRepository<Bill,Long>{
+public interface BillRepository extends JpaRepository<Bill,Integer>{
 }
